refactor(ui): use generated API hook for onboarding submission
